+++ patterns/point.cpp
@@ -1,6 +1,6 @@
 #include "point.h"
 
-Point::Point(uint16_t x, uint16_t y) {
+Point::Point(int32_t x, int32_t y) {
 	this->x = x;
 	this->y = y;
 }
