+++ patterns/point.h
@@ -4,9 +4,9 @@
 #include "FastLED.h"
 
 struct Point {
-	uint16_t x = 0;
-	uint16_t y = 0;
-	Point(uint16_t x, uint16_t y);
+	int32_t x = 0;
+	int32_t y = 0;
+	Point(int32_t x, int32_t y);
 };
 
 #endif
