+++ patterns/tools.h
@@ -5,7 +5,7 @@
 
 class Tools {
   public:
-    static bool in_bounds(int16_t x, int16_t y, uint16_t width, uint16_t height);
+    static bool in_bounds(int32_t x, int32_t y, int32_t width, int32_t height);
 };
 
 #endif
